+++ src/Debug/ConnectionInterfaceProxy.php
@@ -27,7 +27,7 @@ final class ConnectionInterfaceProxy implements ConnectionInterface
     /**
      * @psalm-suppress PossiblyUndefinedArrayOffset
      */
-    public function beginTransaction(string $isolationLevel = null): TransactionInterface
+    public function beginTransaction(?string $isolationLevel = null): TransactionInterface
     {
         [$callStack] = debug_backtrace();
 
@@ -42,7 +42,7 @@ final class ConnectionInterfaceProxy implements ConnectionInterface
         return $this->connection->createBatchQueryResult($query, $each);
     }
 
-    public function createCommand(string $sql = null, array $params = []): CommandInterface
+    public function createCommand(?string $sql = null, array $params = []): CommandInterface
     {
         return new CommandInterfaceProxy(
             $this->connection->createCommand($sql, $params),
@@ -63,7 +63,7 @@ final class ConnectionInterfaceProxy implements ConnectionInterface
         $this->connection->close();
     }
 
-    public function getLastInsertID(string $sequenceName = null): string
+    public function getLastInsertID(?string $sequenceName = null): string
     {
         return $this->connection->getLastInsertID($sequenceName);
     }
@@ -144,7 +144,7 @@ final class ConnectionInterfaceProxy implements ConnectionInterface
      * @psalm-param Closure(self): mixed $closure
      * @psalm-suppress PossiblyUndefinedArrayOffset
      */
-    public function transaction(Closure $closure, string $isolationLevel = null): mixed
+    public function transaction(Closure $closure, ?string $isolationLevel = null): mixed
     {
         [$callStack] = debug_backtrace();