+++ src/QueryBuilder/ColumnDefinitionBuilderInterface.php
@@ -21,4 +21,11 @@ interface ColumnDefinitionBuilderInterface
      * Builds column definition for `ALTER` operation based on given column instance.
      */
     public function buildAlter(ColumnInterface $column): string;
+
+    /**
+     * Builds the type definition for the column. For example: `varchar(128)` or `decimal(10,2)`.
+     *
+     * @return string A string containing the column type definition.
+     */
+    public function buildType(ColumnInterface $column): string;
 }