+++ tests/Db/Command/CommandTest.php
@@ -117,18 +117,20 @@ final class CommandTest extends AbstractCommandTest
      * @dataProvider \Yiisoft\Db\Tests\Provider\CommandProvider::addForeignKeySql
      */
     public function testAddForeignKeySql(
-        string $name,
-        string $tableName,
-        array|string $column1,
-        array|string $column2,
+        array|string $columns,
+        array|string $referenceColumns,
         string|null $delete,
         string|null $update,
         string $expected
     ): void {
         $db = $this->getConnection();
-
         $command = $db->createCommand();
-        $sql = $command->addForeignKey($tableName, $name, $column1, $tableName, $column2, $delete, $update)->getSql();
+
+        $name = '{{fk_constraint}}';
+        $tableName = '{{fk_table}}';
+        $referenceTable = '{{fk_referenced_table}}';
+
+        $sql = $command->addForeignKey($tableName, $name, $columns, $referenceTable, $referenceColumns, $delete, $update)->getSql();
 
         $this->assertSame($expected, $sql);
     }
@@ -720,7 +722,7 @@ final class CommandTest extends AbstractCommandTest
         $command->upsert('{{table}}', []);
     }
 
-    public function testProfiler(string $sql = null): void
+    public function testProfiler(?string $sql = null): void
     {
         $this->expectExceptionMessage(
             'Yiisoft\Db\Tests\Support\Stub\Command::internalExecute is not supported by this DBMS.'
@@ -728,7 +730,7 @@ final class CommandTest extends AbstractCommandTest
         parent::testProfiler();
     }
 
-    public function testProfilerData(string $sql = null): void
+    public function testProfilerData(?string $sql = null): void
     {
         $this->expectExceptionMessage(
             'Yiisoft\Db\Tests\Support\Stub\Command::internalExecute is not supported by this DBMS.'