+++ tests/Provider/ColumnProvider.php
@@ -4,26 +4,34 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Tests\Provider;
 
+use ArrayIterator;
 use PDO;
 use stdClass;
 use Yiisoft\Db\Command\Param;
 use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\Constraint\ForeignKeyConstraint;
+use Yiisoft\Db\Expression\ArrayExpression;
 use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Expression\JsonExpression;
 use Yiisoft\Db\Constant\PhpType;
 use Yiisoft\Db\Expression\StructuredExpression;
 use Yiisoft\Db\Schema\Column\ArrayColumn;
+use Yiisoft\Db\Schema\Column\ArrayLazyColumn;
 use Yiisoft\Db\Schema\Column\BigIntColumn;
 use Yiisoft\Db\Schema\Column\BinaryColumn;
 use Yiisoft\Db\Schema\Column\BitColumn;
 use Yiisoft\Db\Schema\Column\BooleanColumn;
-use Yiisoft\Db\Schema\Column\ColumnBuilder;
 use Yiisoft\Db\Schema\Column\DoubleColumn;
 use Yiisoft\Db\Schema\Column\IntegerColumn;
 use Yiisoft\Db\Schema\Column\JsonColumn;
+use Yiisoft\Db\Schema\Column\JsonLazyColumn;
 use Yiisoft\Db\Schema\Column\StringColumn;
 use Yiisoft\Db\Schema\Column\StructuredColumn;
+use Yiisoft\Db\Schema\Column\StructuredLazyColumn;
+use Yiisoft\Db\Schema\Data\LazyArray;
+use Yiisoft\Db\Schema\Data\JsonLazyArray;
+
+use Yiisoft\Db\Schema\Data\StructuredLazyArray;
 
 use function fopen;
 
@@ -50,7 +58,7 @@ class ColumnProvider
     {
         return [
             'integer' => [
-                IntegerColumn::class,
+                new IntegerColumn(),
                 [
                     // [expected, typecast value]
                     [null, null],
@@ -64,7 +72,7 @@ class ColumnProvider
                 ],
             ],
             'bigint' => [
-                BigIntColumn::class,
+                new BigIntColumn(),
                 [
                     [null, null],
                     [null, ''],
@@ -78,7 +86,7 @@ class ColumnProvider
                 ],
             ],
             'double' => [
-                DoubleColumn::class,
+                new DoubleColumn(),
                 [
                     [null, null],
                     [null, ''],
@@ -91,7 +99,7 @@ class ColumnProvider
                 ],
             ],
             'string' => [
-                StringColumn::class,
+                new StringColumn(),
                 [
                     [null, null],
                     ['', ''],
@@ -104,7 +112,7 @@ class ColumnProvider
                 ],
             ],
             'binary' => [
-                BinaryColumn::class,
+                new BinaryColumn(),
                 [
                     [null, null],
                     ['1', 1],
@@ -116,7 +124,7 @@ class ColumnProvider
                 ],
             ],
             'bit' => [
-                BitColumn::class,
+                new BitColumn(),
                 [
                     [null, null],
                     [null, ''],
@@ -131,7 +139,7 @@ class ColumnProvider
                 ],
             ],
             'boolean' => [
-                BooleanColumn::class,
+                new BooleanColumn(),
                 [
                     [null, null],
                     [null, ''],
@@ -147,7 +155,7 @@ class ColumnProvider
                 ],
             ],
             'json' => [
-                JsonColumn::class,
+                new JsonColumn(),
                 [
                     [null, null],
                     [new JsonExpression(''), ''],
@@ -160,6 +168,29 @@ class ColumnProvider
                     [new JsonExpression(['a' => 1]), ['a' => 1]],
                     [new JsonExpression(new stdClass()), new stdClass()],
                     [$expression = new JsonExpression([1, 2, 3]), $expression],
+                    [$expression = new Expression('expression'), $expression],
+                ],
+            ],
+            'array' => [
+                $arrayCol = new ArrayColumn(),
+                [
+                    [null, null],
+                    [new ArrayExpression([], $arrayCol), []],
+                    [new ArrayExpression([1, 2, 3], $arrayCol), [1, 2, 3]],
+                    [new ArrayExpression($iterator = new ArrayIterator([1, 2, 3]), $arrayCol), $iterator],
+                    [new ArrayExpression('[1,2,3]', $arrayCol), '[1,2,3]'],
+                    [$expression = new Expression('expression'), $expression],
+                ],
+            ],
+            'structured' => [
+                $structuredCol = new StructuredColumn(),
+                [
+                    [null, null],
+                    [new StructuredExpression([], $structuredCol), []],
+                    [new StructuredExpression(['value' => 1, 'currency_code' => 'USD'], $structuredCol), ['value' => 1, 'currency_code' => 'USD']],
+                    [new StructuredExpression($iterator = new ArrayIterator(['value' => 1, 'currency_code' => 'USD']), $structuredCol), $iterator],
+                    [new StructuredExpression('[1,"USD"]', $structuredCol), '[1,"USD"]'],
+                    [$expression = new Expression('expression'), $expression],
                 ],
             ],
         ];
@@ -169,7 +200,7 @@ class ColumnProvider
     {
         return [
             'integer' => [
-                IntegerColumn::class,
+                new IntegerColumn(),
                 [
                     // [expected, typecast value]
                     [null, null],
@@ -178,7 +209,7 @@ class ColumnProvider
                 ],
             ],
             'bigint' => [
-                BigIntColumn::class,
+                new BigIntColumn(),
                 [
                     [null, null],
                     ['1', 1],
@@ -187,7 +218,7 @@ class ColumnProvider
                 ],
             ],
             'double' => [
-                DoubleColumn::class,
+                new DoubleColumn(),
                 [
                     [null, null],
                     [1.0, 1.0],
@@ -195,7 +226,7 @@ class ColumnProvider
                 ],
             ],
             'string' => [
-                StringColumn::class,
+                new StringColumn(),
                 [
                     [null, null],
                     ['', ''],
@@ -203,7 +234,7 @@ class ColumnProvider
                 ],
             ],
             'binary' => [
-                BinaryColumn::class,
+                new BinaryColumn(),
                 [
                     [null, null],
                     ['', ''],
@@ -212,7 +243,7 @@ class ColumnProvider
                 ],
             ],
             'bit' => [
-                BitColumn::class,
+                new BitColumn(),
                 [
                     [null, null],
                     [1, 1],
@@ -222,7 +253,7 @@ class ColumnProvider
                 ],
             ],
             'boolean' => [
-                BooleanColumn::class,
+                new BooleanColumn(),
                 [
                     [null, null],
                     [true, true],
@@ -233,7 +264,7 @@ class ColumnProvider
                 ],
             ],
             'json' => [
-                JsonColumn::class,
+                new JsonColumn(),
                 [
                     [null, null],
                     [null, 'null'],
@@ -248,6 +279,61 @@ class ColumnProvider
                     [['a' => 1], '{"a":1}'],
                 ],
             ],
+            'jsonLazy' => [
+                new JsonLazyColumn(),
+                [
+                    [null, null],
+                    [null, 'null'],
+                    ['', '""'],
+                    [1.0, '1.0'],
+                    [1, '1'],
+                    [true, 'true'],
+                    [false, 'false'],
+                    ['string', '"string"'],
+                    [new JsonLazyArray('[1,2,3]'), '[1,2,3]'],
+                    [new JsonLazyArray('{"key":"value"}'), '{"key":"value"}'],
+                ],
+            ],
+            'array' => [
+                new ArrayColumn(),
+                [
+                    [null, null],
+                    [[], '[]'],
+                    [[], '{}'],
+                    [[1, 2, 3], '[1,2,3]'],
+                    [['1', '2', '3'], '["1","2","3"]'],
+                    [['key' => 'value'], '{"key":"value"}'],
+                ],
+            ],
+            'arrayLazy' => [
+                new ArrayLazyColumn(),
+                [
+                    [null, null],
+                    [new LazyArray('[]'), '[]'],
+                    [new LazyArray('{}'), '{}'],
+                    [new LazyArray('[1,2,3]'), '[1,2,3]'],
+                ],
+            ],
+            'structured' => [
+                new StructuredColumn(),
+                [
+                    [null, null],
+                    [[], '[]'],
+                    [[], '{}'],
+                    [[1, true], '[1,true]'],
+                    [['key' => 'value'], '{"key":"value"}'],
+                ],
+            ],
+            'structuredLazy' => [
+                $structuredCol = (new StructuredLazyColumn())->columns(['key' => new StringColumn()]),
+                [
+                    [null, null],
+                    [new StructuredLazyArray('[]', $structuredCol->getColumns()), '[]'],
+                    [new StructuredLazyArray('{}', $structuredCol->getColumns()), '{}'],
+                    [new StructuredLazyArray('[1,true]', $structuredCol->getColumns()), '[1,true]'],
+                    [new StructuredLazyArray('{"key":"value"}', $structuredCol->getColumns()), '{"key":"value"}'],
+                ],
+            ],
         ];
     }
 
@@ -256,7 +342,7 @@ class ColumnProvider
         return [
             // [column, values]
             ColumnType::BOOLEAN => [
-                ColumnBuilder::boolean(),
+                new BooleanColumn(),
                 [
                     // [dimension, expected, typecast value]
                     [1, [true, true, true, false, false, false, null], [true, 1, '1', false, 0, '0', null]],
@@ -264,14 +350,14 @@ class ColumnProvider
                 ],
             ],
             ColumnType::BIT => [
-                ColumnBuilder::bit(),
+                new BitColumn(),
                 [
                     [1, [0b1011, 1001, null], [0b1011, '1001', null]],
                     [2, [[0b1011, 1001, null]], [[0b1011, '1001', null]]],
                 ],
             ],
             ColumnType::INTEGER => [
-                ColumnBuilder::integer(),
+                new IntegerColumn(),
                 [
                     [1, [1, 2, 3, null], [1, 2.0, '3', null]],
                     [2, [[1, 2], [3], null], [[1, 2.0], ['3'], null]],
@@ -286,21 +372,21 @@ class ColumnProvider
                 ],
             ],
             ColumnType::DOUBLE => [
-                ColumnBuilder::double(),
+                new DoubleColumn(),
                 [
                     [1, [1.0, 2.2, 3.3, null], [1, 2.2, '3.3', null]],
                     [2, [[1.0, 2.2], [3.3, null]], [[1, 2.2], ['3.3', null]]],
                 ],
             ],
             ColumnType::STRING => [
-                ColumnBuilder::string(),
+                new StringColumn(),
                 [
                     [1, ['1', '2', '1', '0', '', null], [1, '2', true, false, '', null]],
                     [2, [['1', '2', '1', '0'], [''], [null]], [[1, '2', true, false], [''], [null]]],
                 ],
             ],
             ColumnType::BINARY => [
-                ColumnBuilder::binary(),
+                new BinaryColumn(),
                 [
                     [1, [
                         '1',
@@ -317,7 +403,7 @@ class ColumnProvider
                 ],
             ],
             ColumnType::JSON => [
-                ColumnBuilder::json(),
+                new JsonColumn(),
                 [
                     [1, [
                         new JsonExpression([1, 2, 3]),
@@ -337,7 +423,7 @@ class ColumnProvider
                 ],
             ],
             ColumnType::STRUCTURED => [
-                ColumnBuilder::structured('structured_type'),
+                (new StructuredColumn())->dbType('structured_type'),
                 [
                     [
                         1,