+++ CHANGELOG.md
@@ -55,6 +55,12 @@
 - Enh #917: Rename `ColumnSchemaInterface` to `ColumnInterface` (@Tigrov)
 - Enh #919: Replace `name()` with immutable `withName()` method in `ColumnInterface` interface (@Tigrov)
 - Enh #921: Move `DataType` class to `Yiisoft\Db\Constant` namespace (@Tigrov)
+- Enh #926: Refactor `DbArrayHelper` (@Tigrov)
+- Enh #920: Move index constants to the appropriate DBMS driver's `IndexType` and `IndexMethod` classes (@Tigrov)
+- New #928: Add `ReferentialAction` class with constants of possible values of referential actions (@Tigrov)
+- Enh #929: Refactor array, structured and JSON column type expressions and expression builders (@Tigrov)
+- Enh #929: Implement lazy arrays for array, structured and JSON column types (@Tigrov)
+- Bug #933: Explicitly mark nullable parameters (@vjik)
 
 ## 1.3.0 March 21, 2024